feat: added autoversioning capability to the premake script
During configuration, a prompt is now supplied to determine if auto-versioning should be applied.
If used, an "autoversion.h" file is created in the '/src/' directory with values populated from
the git tag. If no git tag is found/there is an error retrieving the value(s), default values
of 0 are utilized.

If auto-versioning is not used, nothing has changed _except_ for cleaning up some pre-defined fields
in an object in the '/premake/premake5.lua' script which were not actually needed.

To access the values associated with the version, the following macros are supplied (where $PRJ.NAME$
is replaced by the project's name (in caps) with any spaces/hyphens replaced with underscores:

- $PRJ.NAME$_VERSION_STRING
- $PRJ.NAME$_VERSION_MAJOR
- $PRJ.NAME$_VERSION_MINOR
- $PRJ.NAME$_VERSION_PATCH
- $PRJ.NAME$_COMMIT_HASH

For example, a project "Example Project" in the directory '/example project/' would have macros
EXAMPLE_PROJECT_VERSION_STRING and so on.